+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.13.2
+
+* feat: support for `Project` model
+* feat: add `AppProject` to `App` interface
+
 ## 0.13.1
 
 * feat: add `is_limited` to `Collaborator` interface